pyproject.toml
[project]
name = "url-monitor"
version = "0.1.0"
description = "A simple Python script to monitor a URL for changes and send Telegram notifications."
authors = [{ name = "Gemini" }]
requires-python = ">=3.8"
dependencies = [
    "requests",
    "python-telegram-bot",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "coverage",
    "types-requests",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
# Same as Black.
line-length = 88

# Enable Pyflakes `E` and `F` codes by default.
# E501 is line-length, which is handled by the line-length setting.
select = ["E", "F"]
ignore = ["E501"]

# Allow unused variables in `__init__.py`.
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
strict = true
ignore_missing_imports = true


